Standardize InventoryActionEvent to use Event.Result
